requirements.txt
pyyaml
chromadb
sentence-transformers
rank-bm25
duckduckgo-search
ddgs
gradio
